Introspect top-level folder of idr-metadata from the script location
The annotation scripts currently expect the idr-metadata folder to be located
at /tmp/idr-metadata. This assumption is not practical as soon as multiple
users want to run different dataset imports or annotations on the server.

This change allows the annotation script to introspect the location of the
idr-metadata folder, making it easier for anyone to use a fresh clone on an
arbitrary folder to run the annotation scripts.