Remove update_compounds script
Initially the idea of the script was to be able to update IDR compounds
to link to the relevant PubChem URL.
As demonstrated by the following compound studies, the absence of
standardization on compound names certainly makes this complex as:
- same CIDs have name variants
- different compound entities have the same name

In addition, an incorrect CID mapping would currently update all the compounds
of the same name. There is intrinsic danger with this tight coupling.

Instead the approach chosen for prod90 is to define IDR compounds as a combination of name + URL and keep the annotation/re-annotation per study
using the metadata submitted and curated.