Add script to clean orphaned map annotations
Due to scaling limitations with the graph operations, the deletion of primary
map annotations can leave orphaned map annotations in the IDR database.

This script loops over the primary namespaces, queries all map annotations
attached neither to an image nor to a well and performs a delete operation on
the results using batches of 500 by default.